requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
cv==1.0.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
